+++ backend/api/database.js
@@ -1,0 +1,36 @@
+var mysql = require('mysql');
+
+class Database {
+  constructor(config) {
+    this.config = config;
+    this.pool = mysql.createPool(this.config);
+  }
+
+  query(sql, args) {
+    return new Promise((resolve, reject) => {
+      this.pool.getConnection(function(err, connection) {
+        if (err) {
+          console.log("Error getting a MySQL connection from pool. ", err);
+          reject(err);
+        }
+
+        connection.query(sql, args, (err, rows) => {
+          if (err) reject(err);
+          else resolve(rows);
+
+          connection.release();
+        });
+      });
+    });
+  }
+
+  connection() {
+    return this.pool.getConnection;
+  }
+
+  close() {
+    this.pool.end();
+  }
+}
+
+module.exports = Database;
