+++ backend/api/locale.js
@@ -1,5 +1,26 @@
 
-const strings = {
+class Locale {
+
+  static apply(data, locale) {
+
+    if (data[locale] !== undefined) {
+      return data[locale];
+    }
+    else {
+      for (var key in data) {
+        if (data[key] instanceof Object) {
+          data[key] = _localize(data[key], locale);
+        }
+        else if (data[key] instanceof Array) {
+          data[key] = data[key].map(_ => _localize(_, locale));
+        }
+      }
+      return data;
+    }
+  }
+}
+
+Locale.strings = {
   sub_type_string: {
     SelfEmployedBusiness: {
       SOCIAL_MEDIA_SPECIALIST: {
@@ -330,4 +351,4 @@ const strings = {
   }
 }
 
-module.exports = strings;
+module.exports = Locale;
